// src/types/wine.ts
export interface Wine {
    id: number;
    Location: string;
    Rating: string;
    Vintage: string;
    Winery: string;
    Grape: string;
    Style: string;
    Name: string;
    Region: string;
    Helper: string;
    Price: string;
    Consumed: string;
    Date_Consumed: string;
    Notes: string;
}